requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
